infrastructure/requirements.txt
aws-cdk-lib>=2.160,<3
constructs>=10.3,<11
